[FEAT] Move sw update logic to hook
