+++ src/features/island/hud/components/BumpkinProfile.tsx
@@ -177,7 +177,7 @@ export const BumpkinAvatar: React.FC<AvatarProps> = ({
         >
           {level}
         </div>
-        {true && (
+        {username && (
           <div
             className={`col-start-1 row-start-1 flex justify-center text-white text-xxs z-20`}
             style={{
@@ -187,7 +187,7 @@ export const BumpkinAvatar: React.FC<AvatarProps> = ({
               marginTop: `${DIMENSIONS.username.marginTop}px`,
             }}
           >
-            {`Craig Boss`}
+            {username}
           </div>
         )}
         {showSkillPointAlert && (
