+++ src/lib/utils/hooks/useServiceWorkerUpdate.ts
@@ -1,0 +1,56 @@
+/* eslint-disable no-console */
+import { useEffect, useRef } from "react";
+import { useRegisterSW } from "virtual:pwa-register/react";
+import lifecycle from "page-lifecycle/dist/lifecycle.mjs";
+
+const CHECK_FOR_UPDATE_INTERVAL = 1000 * 60 * 4;
+
+export function useServiceWorkerUpdate() {
+  const {
+    needRefresh: [needRefresh],
+    updateServiceWorker,
+  } = useRegisterSW({
+    onRegisteredSW(swUrl, registration) {
+      if (registration) {
+        setInterval(async () => {
+          if (!(!registration.installing && navigator)) return;
+
+          if ("connection" in navigator && !navigator.onLine) return;
+
+          const resp = await fetch(swUrl, {
+            cache: "no-store",
+            headers: {
+              cache: "no-store",
+              "cache-control": "no-cache",
+            },
+          });
+
+          if (resp?.status === 200) await registration.update();
+        }, CHECK_FOR_UPDATE_INTERVAL);
+      }
+    },
+  });
+
+  const needRefreshRef = useRef(needRefresh);
+
+  useEffect(() => {
+    needRefreshRef.current = needRefresh;
+  }, [needRefresh]);
+
+  useEffect(() => {
+    const handleStateChange = (evt: any) => {
+      console.log("State change: ", evt.newState);
+      console.log("Need refresh: ", needRefreshRef.current);
+      if (evt.newState === "hidden" && needRefreshRef.current) {
+        updateServiceWorker();
+      }
+    };
+
+    lifecycle.addEventListener("statechange", handleStateChange);
+
+    return () => {
+      lifecycle.removeEventListener("statechange", handleStateChange);
+    };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+}
